Enable monthly updating from the site
Bring back the automatic updating of records that are at least a month old. When such a record is being requested by a user, it is fetched from the site instead of the DB, updating the latter. In case of an error, fallback to the old database record.

Block name input until page init.